Workflow: serialize package-level metadata before backlog
This commit ensures that "Serialize Dublin Core metadata to disk" is also used
in the "Send to backlog" workflow so transfers in backlog maintain a copy of
`dc.json` (package-level metadata).

The goal is to ensure that the metadata is persisted in the transfer package.
However, Archivematica is still unable to consolidate package-level metadata
found in `dc.json` during SIP arrangement. That will require further work.